Change in workflow, read detailed description below.
a. Deleted the index.html and css from the root folder (PictureThis).
b. The workflow from now onwards is:
    - Before deplyoying to Google App Engine, copy the index.html and
css folders from "PicThisGap>>www" and paste it into the root
PictureThis folder.

We did this because the native application always has more
functionality, and it made sense to make that the parent copy.